eldk-switch.sh: Fix ARCH setting in ELDKs >= 5.4
Beginning with version 5.4 the environment-setup script contains a setting
for ARCH.  As we evaluate all variables simply to get TARGET_PREFIX this
overwrites ARCH from the calling shell.  Subsequently the ARCH logic
(needed for the transition phase where two ARCH settings were valid,
i.e. "ppc" and "powerpc") fails and does not output commands for
setting ARCH.

The proper fix is to only evaluate what we really need in the script.

Signed-off-by: Detlev Zundel <dzu@denx.de>